casts and defines to make solaris happy

+++ source/utils/nbio.c
@@ -42,7 +42,7 @@ static void sigsegv(int sig)
 	char line[200];
 	printf("segv at line %d\n", line_count);
 	slprintf(line, sizeof(line), "/usr/X11R6/bin/xterm -e gdb /proc/%d/exe %d", 
-		getpid(), getpid());
+		(int)getpid(), (int)getpid());
 	system(line);
 	exit(1);
 }
