+++ source/utils/torture.c
@@ -1416,7 +1416,7 @@ static void run_dirtest(int dummy)
 	srandom(0);
 	for (i=0;i<numops;i++) {
 		fstring fname;
-		slprintf(fname, sizeof(fname), "%x", random());
+		slprintf(fname, sizeof(fname), "%x", (int)random());
 		fnum = cli_open(&cli, fname, O_RDWR|O_CREAT, DENY_NONE);
 		if (fnum == -1) {
 			fprintf(stderr,"Failed to open %s\n", fname);
@@ -1436,7 +1436,7 @@ static void run_dirtest(int dummy)
 	srandom(0);
 	for (i=0;i<numops;i++) {
 		fstring fname;
-		slprintf(fname, sizeof(fname), "%x", random());
+		slprintf(fname, sizeof(fname), "%x", (int)random());
 		cli_unlink(&cli, fname);
 	}
